Rename allure-specflow projects and fix tests.
Tests of the Release version of Allure.SpecFlow incorrectly invoked
the Debug configuration of samples.
