+++ Allure.SpecFlow.Tests.Samples/AssemblyInfo.cs
