+++ Allure.SpecFlow.Tests/Allure.SpecFlow.Tests.csproj
@@ -11,7 +11,7 @@
     </ItemGroup>
 
     <ItemGroup>
-        <ProjectReference Include="..\Allure.SpecFlowPlugin\Allure.SpecFlowPlugin.csproj" />
+        <ProjectReference Include="..\Allure.SpecFlow\Allure.SpecFlow.csproj" />
     </ItemGroup>
 
     <ItemGroup>
