+++ Allure.SpecFlow.Tests/TestSetup.cs
@@ -1,0 +1,17 @@
+﻿using NUnit.Framework;
+using System;
+using System.IO;
+
+namespace Allure.SpecFlow.Tests
+{
+    [SetUpFixture]
+    public class TestSetup
+    {
+        [OneTimeSetUp]
+        public void Setup()
+        {
+            // setup current folder for nUnit engine
+            Environment.CurrentDirectory = Path.GetDirectoryName(typeof(TestSetup).Assembly.Location);
+        }
+    }
+}
