+++ Allure.SpecFlow.Tests/allureConfigWithInvalidRegex.json
