+++ Allure.SpecFlow/SelectiveRun/AllureSpecFlowPatcher.cs
