--- Allure.SpecFlowPlugin.Tests/TestSetup.cs
@@ -1,17 +1,0 @@
-﻿using NUnit.Framework;
-using System;
-using System.IO;
-
-namespace Allure.SpecFlowPlugin.Tests
-{
-  [SetUpFixture]
-  public class TestSetup
-  {
-    [OneTimeSetUp]
-    public void Setup()
-    {
-      // setup current folder for nUnit engine
-      Environment.CurrentDirectory = Path.GetDirectoryName(typeof(TestSetup).Assembly.Location);
-    }
-  }
-}
